moved routes from evaluatoor to core v2
