+++ story_manager/story_manager/main.py
@@ -7,9 +7,9 @@ from commons.utilities.docs import custom_openapi, setup_swagger_ui
 from commons.utilities.logger import setup_rich_logger
 from story_manager.config import get_settings
 from story_manager.core.routes import router as core_router
+from story_manager.core.v2.routes import router as stories_v2_router
 from story_manager.exceptions import add_exception_handlers
 from story_manager.health import router as health_check_router
-from story_manager.story_evaluator.routes import router as story_evaluator_router
 
 
 def get_application() -> FastAPI:
@@ -25,7 +25,7 @@ def get_application() -> FastAPI:
     )
     _app.include_router(core_router, prefix="/v1")
     _app.include_router(health_check_router, prefix="/v1")
-    _app.include_router(story_evaluator_router, prefix="/v2")
+    _app.include_router(stories_v2_router, prefix="/v2")
     swagger_router = setup_swagger_ui("Story Manager", settings.URL_PREFIX)
     _app.include_router(swagger_router)
     _app.openapi = custom_openapi(_app, settings)  # type: ignore
